Refactor routes, get audio analysis from Spotify API
* Run nodemon in development instead of node
* Pull routes out of app.js and into routes.js to make it a bit more organized * Update /analyze/:id endpoint to also pull track data
* Get audio analysis from Spotify API
* Add concurrency
