+++ package.json
@@ -5,11 +5,13 @@
   "main": "index.js",
   "scripts": {
     "test": "mocha",
+    "dev": "nodemon app.js",
     "start": "node app.js"
   },
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "bluebird": "^3.5.1",
     "cookie-parser": "^1.4.3",
     "cors": "^2.8.4",
     "dotenv": "^4.0.0",
@@ -19,6 +21,7 @@
     "ms": "^2.0.0",
     "node-localstorage": "^1.3.0",
     "node-pre-gyp": "^0.6.36",
+    "nodemon": "^1.12.1",
     "nopt": "^4.0.1",
     "passport": "^0.4.0",
     "passport-oauth": "^1.0.0",
@@ -27,6 +30,7 @@
     "qs": "^6.5.0",
     "readable-stream": "^2.3.3",
     "request": "^2.81.0",
+    "request-promise": "^4.2.2",
     "string_decoder": "^1.0.3"
   },
   "devDependencies": {
